Ensure filename is a directory
This is a minor suggestion to ensure the files listed in os.listdir are directories

This change assumes that the exception raised when requirements.txt is not found is desired.

Without checking if the filename is a directory, it will enforce the structure that no other files are allowed in the same directory.